Refactoring of functions into respective header and cpp files
