+++ VideoSearch/VideoSearch.vcxproj
@@ -91,6 +91,10 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="src\main.cpp" />
+    <ClCompile Include="src\VideoState.cpp" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="src\VideoState.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
